requirements.txt
atomicwrites==1.4.0
attrs==20.2.0
click==7.1.2
colorama==0.4.3
Flask==1.1.2
importlib-metadata==2.0.0
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.8.0
marshmallow-dataclass==8.0.0
mypy-extensions==0.4.3
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.0
pytest-mock==3.3.1
python-dotenv==0.14.0
six==1.15.0
toml==0.10.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
Werkzeug==1.0.1
zipp==3.2.0
